Removed user_queries, now using the Accounts context
