+++ apps/multirotor_web/web/controllers/page_controller.ex
@@ -2,7 +2,7 @@ defmodule MultirotorWeb.PageController do
   use MultirotorWeb.Web, :controller
 
   def index(conn, _params) do
-    post_list = Multirotor.PostQueries.get_top(100)
+    post_list = Multirotor.Posts.get_top(100)
     render conn, "index.html", posts: post_list
   end
