+++ apps/multirotor_web/web/controllers/post_controller.ex
@@ -1,10 +1,11 @@
 defmodule MultirotorWeb.PostController do
   use MultirotorWeb.Web, :controller
 
+
   plug MultirotorWeb.AuthorizedPlug when action in [:create]
 
   def show(conn, %{"id" => id}) do
-    post = Multirotor.PostQueries.get_by_id(id)
+    post = Multirotor.Posts.get_by_id(id)
     changeset = Multirotor.Post.changeset(%Multirotor.Post{}, %{type: "answer"})
     answerList = Multirotor.PostMappingQueries.get_answers(id)
 
@@ -12,7 +13,7 @@ defmodule MultirotorWeb.PostController do
   end
 
   def index(conn, _params) do
-    post_index = Multirotor.PostQueries.get_all
+    post_index = Multirotor.Posts.get_all
     render conn, "list.html", posts: post_index
   end
 
@@ -29,7 +30,7 @@ defmodule MultirotorWeb.PostController do
     current_user = get_session(conn, :current_user)
 
     changeset = Multirotor.Post.map_answer(posts, current_user.id, 1)
-    case Multirotor.PostQueries.create changeset do
+    case Multirotor.Posts.create changeset do
       {:ok, %{id: id}} -> redirect conn, to: post_path(conn, :show, id)
       {:error, reasons} -> new conn, %{errors: reasons}
     end
@@ -40,7 +41,7 @@ defmodule MultirotorWeb.PostController do
 
     changeset = Multirotor.Post.map_answer(posts,current_user.id,2)
     #TODO how to clean up this nested case statment
-    case Multirotor.PostQueries.create changeset do
+    case Multirotor.Posts.create changeset do
       {:ok, %{id: id}} ->
         questionid = Map.get(conn.path_params,"id")
         mappingChangeset = Multirotor.PostMapping.changeset(%Multirotor.PostMapping{}, %{answerid: id, questionid: questionid})
